src/test/java/com/leetcode/strings/easy/FlipGame_293Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.List;
import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class FlipGame_293Test {
  private final FlipGame_293 solution = new FlipGame_293();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments("++++", List.of("--++", "+--+", "++--")));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void findTheDifference(String str, List<String> expectedResult) {
    List<String> actualResult = solution.generatePossibleNextMoves(str);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
